+++ Sources/StreamingCSVMacros/Plugin.swift
@@ -5,6 +5,8 @@ import SwiftSyntaxMacros
 struct StreamingCSVMacrosPlugin: CompilerPlugin {
     let providingMacros: [Macro.Type] = [
         CSVRowBuilderMacro.self,
+        CSVRowDecoderBuilderMacro.self,
+        CSVRowEncoderBuilderMacro.self,
         FieldMacro.self,
         FieldsMacro.self
     ]